Extract ChatMessage component
